requirements.txt
arabic-reshaper==2.1.3
argcomplete==1.10.0
asgiref==3.4.0
beautifulsoup4==4.8.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==8.0.1
cryptography==3.4.7
cuser==0.0.4
defusedxml==0.7.1
Django==3.2.4
django-admin-sortable2==1.0
django-allauth==0.43.0
django-baton==2.1.5
django-ckeditor==6.1.0
django-contact-us==0.4.4
django-cuser==2017.3.16
django-debug-toolbar==3.1.1
django-embed-video==1.3.3
django-error-report==0.2.0
django-extensions==3.1.3
django-isbn-field==0.5.2
django-js-asset==1.2.2
django-maintenance-mode==0.15.0
django-mptt==0.12.0
django-mptt-admin==2.0.2
django-next-prev==1.0.1
django-taggit==1.4.0
django-user-agents==0.4.0
django-userforeignkey==0.3.0
docx-mailmerge==0.5.0
docx2txt==0.8
dropbox==10.5.0
EbookLib==0.17.1
extract-msg==0.23.1
filehash==0.1.dev5
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
IMAPClient==2.1.0
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
mailmerge==2.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
MouseInfo==0.1.3
nltk==3.6.1
oauthlib==3.1.1
olefile==0.46
pdfminer.six==20181108
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
PyGetWindow==0.0.9
PyJWT==2.1.0
PyMsgBox==1.0.9
pyobjc==7.3
pyobjc-core==7.3
pyobjc-framework-Accessibility==7.3
pyobjc-framework-Accounts==7.3
pyobjc-framework-AddressBook==7.3
pyobjc-framework-AdServices==7.3
pyobjc-framework-AdSupport==7.3
pyobjc-framework-AppleScriptKit==7.3
pyobjc-framework-AppleScriptObjC==7.3
pyobjc-framework-ApplicationServices==7.3
pyobjc-framework-AppTrackingTransparency==7.3
pyobjc-framework-AuthenticationServices==7.3
pyobjc-framework-AutomaticAssessmentConfiguration==7.3
pyobjc-framework-Automator==7.3
pyobjc-framework-AVFoundation==7.3
pyobjc-framework-AVKit==7.3
pyobjc-framework-BusinessChat==7.3
pyobjc-framework-CalendarStore==7.3
pyobjc-framework-CallKit==7.3
pyobjc-framework-CFNetwork==7.3
pyobjc-framework-ClassKit==7.3
pyobjc-framework-CloudKit==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Collaboration==7.3
pyobjc-framework-ColorSync==7.3
pyobjc-framework-Contacts==7.3
pyobjc-framework-ContactsUI==7.3
pyobjc-framework-CoreAudio==7.3
pyobjc-framework-CoreAudioKit==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-CoreData==7.3
pyobjc-framework-CoreHaptics==7.3
pyobjc-framework-CoreLocation==7.3
pyobjc-framework-CoreMedia==7.3
pyobjc-framework-CoreMediaIO==7.3
pyobjc-framework-CoreMIDI==7.3
pyobjc-framework-CoreML==7.3
pyobjc-framework-CoreMotion==7.3
pyobjc-framework-CoreServices==7.3
pyobjc-framework-CoreSpotlight==7.3
pyobjc-framework-CoreText==7.3
pyobjc-framework-CoreWLAN==7.3
pyobjc-framework-CryptoTokenKit==7.3
pyobjc-framework-DeviceCheck==7.3
pyobjc-framework-DictionaryServices==7.3
pyobjc-framework-DiscRecording==7.3
pyobjc-framework-DiscRecordingUI==7.3
pyobjc-framework-DiskArbitration==7.3
pyobjc-framework-DVDPlayback==7.3
pyobjc-framework-EventKit==7.3
pyobjc-framework-ExceptionHandling==7.3
pyobjc-framework-ExecutionPolicy==7.3
pyobjc-framework-ExternalAccessory==7.3
pyobjc-framework-FileProvider==7.3
pyobjc-framework-FileProviderUI==7.3
pyobjc-framework-FinderSync==7.3
pyobjc-framework-FSEvents==7.3
pyobjc-framework-GameCenter==7.3
pyobjc-framework-GameController==7.3
pyobjc-framework-GameKit==7.3
pyobjc-framework-GameplayKit==7.3
pyobjc-framework-ImageCaptureCore==7.3
pyobjc-framework-IMServicePlugIn==7.3
pyobjc-framework-InputMethodKit==7.3
pyobjc-framework-InstallerPlugins==7.3
pyobjc-framework-InstantMessage==7.3
pyobjc-framework-Intents==7.3
pyobjc-framework-IOSurface==7.3
pyobjc-framework-iTunesLibrary==7.3
pyobjc-framework-KernelManagement==7.3
pyobjc-framework-LatentSemanticMapping==7.3
pyobjc-framework-LaunchServices==7.3
pyobjc-framework-libdispatch==7.3
pyobjc-framework-LinkPresentation==7.3
pyobjc-framework-LocalAuthentication==7.3
pyobjc-framework-MapKit==7.3
pyobjc-framework-MediaAccessibility==7.3
pyobjc-framework-MediaLibrary==7.3
pyobjc-framework-MediaPlayer==7.3
pyobjc-framework-MediaToolbox==7.3
pyobjc-framework-Metal==7.3
pyobjc-framework-MetalKit==7.3
pyobjc-framework-MetalPerformanceShaders==7.3
pyobjc-framework-MetalPerformanceShadersGraph==7.3
pyobjc-framework-MLCompute==7.3
pyobjc-framework-ModelIO==7.3
pyobjc-framework-MultipeerConnectivity==7.3
pyobjc-framework-NaturalLanguage==7.3
pyobjc-framework-NetFS==7.3
pyobjc-framework-Network==7.3
pyobjc-framework-NetworkExtension==7.3
pyobjc-framework-NotificationCenter==7.3
pyobjc-framework-OpenDirectory==7.3
pyobjc-framework-OSAKit==7.3
pyobjc-framework-OSLog==7.3
pyobjc-framework-PassKit==7.3
pyobjc-framework-PencilKit==7.3
pyobjc-framework-Photos==7.3
pyobjc-framework-PhotosUI==7.3
pyobjc-framework-PreferencePanes==7.3
pyobjc-framework-PushKit==7.3
pyobjc-framework-Quartz==7.3
pyobjc-framework-QuickLookThumbnailing==7.3
pyobjc-framework-ReplayKit==7.3
pyobjc-framework-SafariServices==7.3
pyobjc-framework-SceneKit==7.3
pyobjc-framework-ScreenSaver==7.3
pyobjc-framework-ScreenTime==7.3
pyobjc-framework-ScriptingBridge==7.3
pyobjc-framework-SearchKit==7.3
pyobjc-framework-Security==7.3
pyobjc-framework-SecurityFoundation==7.3
pyobjc-framework-SecurityInterface==7.3
pyobjc-framework-ServiceManagement==7.3
pyobjc-framework-Social==7.3
pyobjc-framework-SoundAnalysis==7.3
pyobjc-framework-Speech==7.3
pyobjc-framework-SpriteKit==7.3
pyobjc-framework-StoreKit==7.3
pyobjc-framework-SyncServices==7.3
pyobjc-framework-SystemConfiguration==7.3
pyobjc-framework-SystemExtensions==7.3
pyobjc-framework-UniformTypeIdentifiers==7.3
pyobjc-framework-UserNotifications==7.3
pyobjc-framework-UserNotificationsUI==7.3
pyobjc-framework-VideoSubscriberAccount==7.3
pyobjc-framework-VideoToolbox==7.3
pyobjc-framework-Virtualization==7.3
pyobjc-framework-Vision==7.3
pyobjc-framework-WebKit==7.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyserial==3.5
PySimpleGUI==4.39.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-pptx==0.6.18
python-stdnum==1.16
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2021.1
regex==2021.4.4
reportlab==3.5.68
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
rubicon-objc==0.4.0
sh==1.14.2
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
textract==1.6.3
tqdm==4.61.1
tzlocal==1.5.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.11
user-agents==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==1.4.3
